Split out loginUserWithTwoFactorVerificationCode
